Feature/Tqxf2Xote

Approved-by: Carlos Xavier <u.carlosxis@gmail.com>

+++ src/main/java/br/com/ottimizza/dashboard/repositories/company/CompanyRepository.java
@@ -7,8 +7,9 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepository extends JpaRepository<Company, Long>, CompanyRepositoryCustom {
+public interface CompanyRepository extends JpaRepository<Company, BigInteger>, CompanyRepositoryCustom {
 
+	
 	Optional<Company> findById(BigInteger idCompany);
     
 }
